refactor: rename floder  placement to inner
