+++ scripts/place-ig-test.sh
@@ -14,10 +14,10 @@
 
 
 start_server(){
-    nohup cargo run --package cmd --bin placement-center -- --conf=example/mqtt-cluster/placement-center/node-1.toml 2>/tmp/1.log &
+    nohup cargo run --package cmd --bin inner-center -- --conf=example/mqtt-cluster/inner-center/node-1.toml 2>/tmp/1.log &
     sleep 3
 
-    no1=`ps -ef | grep placement-center  | grep node-1 | grep -v grep | awk '{print $2}'`
+    no1=`ps -ef | grep inner-center  | grep node-1 | grep -v grep | awk '{print $2}'`
     if [ -n "$no1" ]
     then
         echo "placement-center node 1 started successfully. process no: $no1"
@@ -25,7 +25,7 @@ start_server(){
 }
 
 stop_server(){
-    no1=`ps -ef | grep placement-center  | grep node-1 | grep -v grep | awk '{print $2}'`
+    no1=`ps -ef | grep inner-center  | grep node-1 | grep -v grep | awk '{print $2}'`
     if [ -n "$no1" ]
     then
         echo "kill placement center $no1"
@@ -33,13 +33,13 @@ stop_server(){
     fi
 }
 # Clean up
-rm -rf /tmp/robust-test/placement-center*
+rm -rf /tmp/robust-test/inner-center*
 
 # Start Server
 start_server
 
 # Run Placement integration Test
-cargo nextest run --package grpc-clients --test mod -- placement
+cargo nextest run --package grpc-clients --test mod -- inner
 cargo nextest run --package robustmq-test --test mod -- place_server
 
 # Stop Server
