+++ src/grpc-clients/src/mqtt/mod.rs
@@ -65,7 +65,7 @@ pub enum MqttBrokerPlacementReply {
 }
 
 pub mod admin;
-pub mod placement;
+pub mod inner;
 
 async fn call_once(
     client_pool: &ClientPool,
