+++ src/grpc-clients/src/pool.rs
@@ -21,7 +21,7 @@ use mobc::{Connection, Pool};
 use crate::journal::admin::JournalAdminServiceManager;
 use crate::journal::inner::JournalInnerServiceManager;
 use crate::mqtt::admin::MqttBrokerAdminServiceManager;
-use crate::mqtt::placement::MqttBrokerPlacementServiceManager;
+use crate::mqtt::inner::MqttBrokerPlacementServiceManager;
 use crate::placement::journal::JournalServiceManager;
 use crate::placement::kv::KvServiceManager;
 use crate::placement::mqtt::MqttServiceManager;
